Refactor: move routes config to avoid import alias
