--- src/components/config/routes.ts
@@ -1,13 +1,0 @@
-import { RouteLocation } from './types'
-
-export const paths = {
-  home: '/',
-  details: '/details',
-  about: '/about',
-  help: '/help',
-  table: '/table',
-  local: '/local',
-  explore: '/Explore',
-} as {
-  [key: string]: RouteLocation
-}
