+++ src/components/details/MoreLikeThis.tsx
@@ -4,7 +4,7 @@ import { Popover } from '@material-ui/core'
 import { BiMapPin } from 'react-icons/bi'
 import { IoIosPeople } from 'react-icons/io'
 
-import { paths as routes } from 'components/config/routes'
+import { routes } from 'components/config/api'
 import { SwatchOnly } from 'components/legend'
 import { Chip, NeighborhoodList } from 'components/details'
 import { DialogCloseBtn } from 'components/generic/modals'
