+++ src/components/map/hooks.tsx
@@ -17,7 +17,7 @@ import { layerSymbFields } from 'components/legend/config'
 import { useAirtable } from 'components/explore/hooks'
 import { useLocation, useRouteMatch } from 'react-router-dom'
 import { AIRTABLE_CENSUS_BASE } from 'components/config'
-import { paths as routes } from 'components/config/routes'
+import { routes } from 'components/config/api'
 import { useWindowResize } from '../../utils'
 import { iconStyleOverride, POINT_ZOOM_LEVEL } from './config'
 import { flyToPoint, flyToBounds } from './utils'
