+++ src/components/results/ResultsModal.tsx
@@ -5,9 +5,9 @@ import { Dialog } from '@material-ui/core'
 import { DialogCloseBtn, SlideUp } from 'components/generic/modals'
 import { DetailsSchema, GlobalContext } from 'components/context'
 import { useAirtable } from 'components/explore/hooks'
+import { routes } from 'components/config/api'
 import { useStyles } from './styles'
 import { ResultsTable } from './ResultsTable'
-import { paths as routes } from '../config/routes'
 import { LocWithState } from '../config/types'
 import { whittleLangFeats } from './utils'
